requirements.txt
-i https://pypi.org/simple
pylzma==0.5.0
pyqt5-sip==4.19.17
pyqt5==5.12.2
pyxdg==0.26
pyyaml==5.1
